+++ rl_zoo3/enjoy.py
@@ -1,32 +1,23 @@
-<<<<<<< HEAD
 import argparse
 import importlib
 import os
 import sys
-import time
 
-import gym
 import numpy as np
 import torch as th
-from torch import einsum
 import yaml
 from huggingface_sb3 import EnvironmentName
+from stable_baselines3.common.callbacks import tqdm
 from stable_baselines3.common.utils import set_random_seed
-from stable_baselines3.common.vec_env.vec_monitor import VecMonitor
 
-import matplotlib.pyplot as plt
-from matplotlib.pyplot import MultipleLocator
-import seaborn as sns
-
-import utils.import_envs  # noqa: F401 pylint: disable=unused-import
-from utils import ALGOS, create_test_env, get_saved_hyperparams
-from utils.exp_manager import ExperimentManager
-from utils.load_from_hub import download_from_hub
-from utils.utils import StoreDict, get_model_path
-from braincog.model_zoo.base_module import BaseModule
+import rl_zoo3.import_envs  # noqa: F401 pylint: disable=unused-import
+from rl_zoo3 import ALGOS, create_test_env, get_saved_hyperparams
+from rl_zoo3.exp_manager import ExperimentManager
+from rl_zoo3.load_from_hub import download_from_hub
+from rl_zoo3.utils import StoreDict, get_model_path
 
 
-def main():  # noqa: C901
+def enjoy():  # noqa: C901
     parser = argparse.ArgumentParser()
     parser.add_argument("--env", help="environment ID", type=EnvironmentName, default="CartPole-v1")
     parser.add_argument("-f", "--folder", help="Log folder", type=str, default="rl-trained-agents")
@@ -48,7 +39,7 @@ def main():  # noqa: C901
         "--load-checkpoint",
         type=int,
         help="Load checkpoint instead of last model if available, "
-             "you must pass the number of timesteps corresponding to it",
+        "you must pass the number of timesteps corresponding to it",
     )
     parser.add_argument(
         "--load-last-checkpoint",
@@ -58,8 +49,7 @@ def main():  # noqa: C901
     )
     parser.add_argument("--stochastic", action="store_true", default=False, help="Use stochastic actions")
     parser.add_argument(
-        "--norm-reward", action="store_true", default=False,
-        help="Normalize reward if applicable (trained with VecNormalize)"
+        "--norm-reward", action="store_true", default=False, help="Normalize reward if applicable (trained with VecNormalize)"
     )
     parser.add_argument("--seed", help="Random generator seed", type=int, default=0)
     parser.add_argument("--reward-log", help="Where to log reward", default="", type=str)
@@ -71,17 +61,18 @@ def main():  # noqa: C901
         help="Additional external Gym environment package modules to import (e.g. gym_minigrid)",
     )
     parser.add_argument(
-        "--env-kwargs", type=str, nargs="+", action=StoreDict,
-        help="Optional keyword argument to pass to the env constructor"
+        "--env-kwargs", type=str, nargs="+", action=StoreDict, help="Optional keyword argument to pass to the env constructor"
     )
     parser.add_argument(
         "--custom-objects", action="store_true", default=False, help="Use custom objects to solve loading issues"
     )
-
     parser.add_argument(
-        '--model-trace', type=str, default=[], nargs="+",
+        "-P",
+        "--progress",
+        action="store_true",
+        default=False,
+        help="if toggled, display a progress bar using tqdm and rich",
     )
-
     args = parser.parse_args()
 
     # Going through custom gym packages to let them register in the global registory
@@ -108,8 +99,7 @@ def main():  # noqa: C901
         if "rl-trained-agents" not in folder:
             raise e
         else:
-            print(
-                "Pretrained model not found, trying to download it from sb3 Huggingface hub: https://huggingface.co/sb3")
+            print("Pretrained model not found, trying to download it from sb3 Huggingface hub: https://huggingface.co/sb3")
             # Auto-download
             download_from_hub(
                 algo=algo,
@@ -176,8 +166,6 @@ def main():  # noqa: C901
         env_kwargs=env_kwargs,
     )
 
-    env = VecMonitor(env, './videos/test.mp4')
-
     kwargs = dict(seed=args.seed)
     if algo in off_policy_algos:
         # Dummy buffer size as we don't need memory to enjoy the trained agent
@@ -202,21 +190,6 @@ def main():  # noqa: C901
 
     model = ALGOS[algo].load(model_path, env=env, custom_objects=custom_objects, device=args.device, **kwargs)
 
-    trace_model = None
-    trace = []
-    if args.model_trace:
-        try:
-            trace_model = model
-            for sub in args.model_trace:
-                trace_model = getattr(trace_model, sub)
-            # trace_model = getattr(model, args.model_trace)
-            assert isinstance(trace_model, BaseModule), print(trace_model, type(trace_model))
-            trace_model.set_requires_fp(True)
-            length = len(trace_model.get_fp())
-            trace = [[] for i in range(length)]
-        except NameError('Cannot find {} in model'.format(args.model_trace)) as e:
-            raise e
-
     obs = env.reset()
 
     # Deterministic by default except for atari games
@@ -230,8 +203,15 @@ def main():  # noqa: C901
     successes = []
     lstm_states = None
     episode_start = np.ones((env.num_envs,), dtype=bool)
+
+    generator = range(args.n_timesteps)
+    if args.progress:
+        if tqdm is None:
+            raise ImportError("Please install tqdm and rich to use the progress bar")
+        generator = tqdm(generator)
+
     try:
-        for _ in range(args.n_timesteps):
+        for _ in generator:
             action, lstm_states = model.predict(
                 obs,
                 state=lstm_states,
@@ -247,11 +227,6 @@ def main():  # noqa: C901
 
             episode_reward += reward[0]
             ep_len += 1
-
-            if trace_model:
-                tr = trace_model.get_fp(temporal_info=True)
-                for i, t in enumerate(tr):
-                    trace[i].append(t)
 
             if args.n_envs == 1:
                 # For atari the return reward is not the atari score
@@ -294,22 +269,8 @@ def main():  # noqa: C901
     if args.verbose > 0 and len(episode_lengths) > 0:
         print(f"Mean episode length: {np.mean(episode_lengths):.2f} +/- {np.std(episode_lengths):.2f}")
 
-    if trace_model:
-        for i in range(len(trace)):
-            trace[i] = th.cat(trace[i], dim=0)
-        trace = th.cat(trace, dim=-1)
-        trace = einsum('tbc -> ct', trace).detach().cpu().numpy()
-        sns.heatmap(trace, cmap="PuRd")
-        plt.xlabel('timestep')
-        plt.ylabel('neuron state')
-        plt.savefig('plot.png', bbox_inches='tight', dpi=1200)
-        plt.show()
-
     env.close()
 
-=======
-from rl_zoo3.enjoy import enjoy
->>>>>>> 1f9cfcfa3b4a12f374aef12d0cbc13d44b3a7674
 
 if __name__ == "__main__":
     enjoy()